refactor: extend viewer and move files
