+++ testcafe/tests/drive/file_sharing_scenario.js
@@ -6,9 +6,9 @@ import {
   setDownloadPath,
   TESTCAFE_DRIVE_URL
 } from '../helpers/utils'
-import PrivateDrivePage from '../pages/drive-model-private'
-import PublicDrivePage from '../pages/drive-model-public'
-import PublicViewerPage from '../pages/drive-viewer-model-public'
+import PrivateDrivePage from '../pages/drive/drive-model-private'
+import PublicDrivePage from '../pages/drive/drive-model-public'
+import PublicViewerPage from '../pages/drive-viewer/drive-viewer-model-public'
 
 let data = require('../helpers/data')
 const privateDrivePage = new PrivateDrivePage()
