+++ testcafe/tests/pages/drive-viewer/drive-viewer-model.js
@@ -1,6 +1,9 @@
 import { t, Selector } from 'testcafe'
-import { getElementWithTestId, isExistingAndVisibile } from '../helpers/utils'
-import DrivePage from '../pages/drive-model'
+import {
+  getElementWithTestId,
+  isExistingAndVisibile
+} from '../../helpers/utils'
+import DrivePage from '../drive/drive-model'
 
 const drivePage = new DrivePage()
 
@@ -103,17 +106,6 @@ export default class Page {
     }
   }
 
-  async downloadWithToolbar() {
-    await t.hover(this.viewerWrapper)
-    await isExistingAndVisibile(
-      this.btnDownloadViewerToolbar,
-      'Download button in toolbar'
-    )
-    await t
-      .setNativeDialogHandler(() => true)
-      .click(this.btnDownloadViewerToolbar)
-  }
-
   //@param {String} screenshotPath : path for screenshots taken in this test
   //@param {string} fileStartName : file to open to start the navigation testing
   //@param {number} numberOfNavigation : the number of file we want to go through during the test.
@@ -128,16 +120,18 @@ export default class Page {
 
     for (let i = startIndex; i < startIndex + numberOfNavigation; i++) {
       await this.navigateToNextFile(i)
-      await t.fixtureCtx.vr.takeScreenshotAndUpload(
-        `${screenshotPath}-${i}-next`
-      )
+      if (t.fixtureCtx.isVR)
+        await t.fixtureCtx.vr.takeScreenshotAndUpload(
+          `${screenshotPath}-${i}-next`
+        )
     }
 
     for (let i = startIndex + numberOfNavigation; i > startIndex; i--) {
       await this.navigateToPrevFile(i)
-      await t.fixtureCtx.vr.takeScreenshotAndUpload(
-        `${screenshotPath}-${i}-prev`
-      )
+      if (t.fixtureCtx.isVR)
+        await t.fixtureCtx.vr.takeScreenshotAndUpload(
+          `${screenshotPath}-${i}-prev`
+        )
     }
     await this.closeViewer({
       exitWithEsc: false
@@ -165,95 +159,88 @@ export default class Page {
     await t.expect(breadcrumbEnd).contains(`${folderName}`)
   }
 
+  //download using the common download button
   async checkCommonViewerDownload(folderName, fileName) {
     await this.openViewerForFile(fileName)
-    await this.downloadWithToolbar()
+    await t.hover(this.viewerWrapper)
+    await isExistingAndVisibile(
+      this.btnDownloadViewerToolbar,
+      'Download button in toolbar'
+    )
+    await t
+      .setNativeDialogHandler(() => true)
+      .click(this.btnDownloadViewerToolbar)
     await this.closeViewer({
       exitWithEsc: true
     })
   }
 
   //Specific check for audioViewer
-  async checkAudioViewer(fileName) {
-    await this.openViewerForFile(fileName)
-    await t.takeScreenshot()
-
+  async checkAudioViewer() {
     await isExistingAndVisibile(this.audioViewer, 'Audio viewer')
     await isExistingAndVisibile(
       this.audioViewerControls,
       'Audio viewer controls'
     )
-    await this.closeViewer({
-      exitWithEsc: false
-    })
+    if (t.fixtureCtx.isVR) {
+      //wait for file to load to get a good screenshots
+      await t.wait(5000)
+      //modify precision to avoid false positive due to loading state
+      t.fixtureCtx.vr.options.compareSettings = {
+        precision: 100 //precision goes from 0 to 255
+      }
+    }
   }
 
   //Specific check for videoViewer
-  async checkVideoViewer(fileName) {
-    await this.openViewerForFile(fileName)
-    await t.takeScreenshot()
-
+  async checkVideoViewer() {
     await isExistingAndVisibile(this.videoViewer, 'Video viewer')
     await isExistingAndVisibile(
       this.videoViewerControls,
       'Video viewer controls'
     )
-    await this.closeViewer({
-      exitWithEsc: false
-    })
+    if (t.fixtureCtx.isVR) {
+      //wait for file to load to get a good screenshots
+      await t.wait(5000)
+      //modify precision to avoid false positive due to loading state
+      t.fixtureCtx.vr.options.compareSettings = {
+        precision: 100 //precision goes from 0 to 255
+      }
+    }
   }
 
   //Specific check for textViewer
-  async checkTextViewer(fileName) {
-    await this.openViewerForFile(fileName)
-    await t.takeScreenshot()
-
+  async checkTextViewer() {
     await isExistingAndVisibile(this.txtViewer, 'text viewer')
     await isExistingAndVisibile(this.txtViewerContent, 'text viewer controls')
-    await this.closeViewer({
-      exitWithEsc: false
-    })
   }
 
   //Specific check for imageViewer
-  async checkImageViewer(fileName) {
-    await this.openViewerForFile(fileName)
-    await t.takeScreenshot()
-
+  async checkImageViewer() {
     await isExistingAndVisibile(this.imageViewer, 'image viewer')
     await isExistingAndVisibile(
       this.imageViewerContent,
       'image viewer controls'
     )
-    await this.closeViewer({
-      exitWithEsc: false
-    })
   }
 
   //Specific check for no viewer : other download btn
-  async checkNoViewer(fileName) {
-    await this.openViewerForFile(fileName)
-    await t.takeScreenshot()
-
+  async checkNoViewer() {
     await isExistingAndVisibile(this.noViewer, 'no-viewer Viewer')
+  }
+
+  //Specific check for no viewer : other download btn
+  async checkNoViewerDownload() {
     await isExistingAndVisibile(
       this.btnNoViewerDownload,
       'no Viewer Download butoon'
     )
     await t.setNativeDialogHandler(() => true).click(this.btnNoViewerDownload)
-
-    await this.closeViewer({
-      exitWithEsc: true
-    })
   }
 
   //Specific check for pdf viewer : download btn
-  async checkPdfViewer(fileName) {
-    await this.openViewerForFile(fileName)
+  async checkPdfViewer() {
     await t.takeScreenshot()
     await isExistingAndVisibile(this.pdfViewer, 'Pdf Viewer')
-    await this.closeViewer({
-      exitWithEsc: true
-    })
   }
 }
