+++ testcafe/tests/pages/drive/drive-model-public.js
@@ -4,8 +4,8 @@ import {
   isExistingAndVisibile,
   getPageUrl,
   goBack
-} from '../helpers/utils'
-import DrivePage from '../pages/drive-model'
+} from '../../helpers/utils'
+import DrivePage from './drive-model'
 
 export default class PublicDrivePage extends DrivePage {
   constructor() {
